Refactor: Move get_user_id function to database.py
- Move get_user_id function into the database.py file
- Remove usernameID.py file
- Adjust imports in routes.py
